+++ model_compression_toolkit/exporter/__init__.py
@@ -15,8 +15,8 @@
 
 from model_compression_toolkit.core.common.constants import FOUND_TF
 
-
 if FOUND_TF:
-    from model_compression_toolkit.exporter.fully_quantized.keras.builder.fully_quantized_model_builder import \
-        get_fully_quantized_keras_model
-    from model_compression_toolkit.exporter.target_platform_export.keras import export_keras_fully_quantized_model
+    from model_compression_toolkit.exporter.model_wrapper.keras.builder.fully_quantized_model_builder import \
+        get_exportable_keras_model
+    from model_compression_toolkit.exporter.model_exporter.keras.keras_export_facade import keras_export_model
+
