+++ model_compression_toolkit/exporter/model_exporter/keras/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from model_compression_toolkit.exporter.target_platform_export.keras.export_keras_fully_quantized_model import \
-    export_keras_fully_quantized_model, KerasExportMode
+from model_compression_toolkit.exporter.model_exporter.keras.keras_export_facade import \
+    keras_export_model, KerasExportMode
