+++ model_compression_toolkit/exporter/model_wrapper/pytorch/builder/node_to_quantize_config.py
@@ -15,19 +15,19 @@
 from model_compression_toolkit.core.common import BaseNode
 from model_compression_toolkit.core.pytorch.back2framework.quantization_wrapper.wrapper_quantize_config import \
     WrapperQuantizeConfig
-from model_compression_toolkit.exporter.fully_quantized.pytorch.builder.node_to_quantizer import \
+from model_compression_toolkit.exporter.model_wrapper.pytorch.builder.node_to_quantizer import \
     get_weights_quantizer_for_node, \
     get_activations_quantizer_for_node
-from model_compression_toolkit.exporter.fully_quantized.pytorch.wrappers_quantize_configs.activation_quantize_config \
+from model_compression_toolkit.exporter.model_wrapper.pytorch.wrappers_quantize_configs.activation_quantize_config \
     import \
     ActivationQuantizeConfig
-from model_compression_toolkit.exporter.fully_quantized.pytorch.wrappers_quantize_configs \
+from model_compression_toolkit.exporter.model_wrapper.pytorch.wrappers_quantize_configs \
     .no_quantization_quantize_config import \
     NoQuantizationQuantizeConfig
-from model_compression_toolkit.exporter.fully_quantized.pytorch.wrappers_quantize_configs \
+from model_compression_toolkit.exporter.model_wrapper.pytorch.wrappers_quantize_configs \
     .weights_activation_quantize_config import \
     WeightsActivationQuantizeConfig
-from model_compression_toolkit.exporter.fully_quantized.pytorch.wrappers_quantize_configs.weights_quantize_config \
+from model_compression_toolkit.exporter.model_wrapper.pytorch.wrappers_quantize_configs.weights_quantize_config \
     import \
     WeightsQuantizeConfig
