+++ model_compression_toolkit/exporter/model_wrapper/pytorch/builder/node_to_quantizer.py
@@ -13,18 +13,20 @@
 # limitations under the License.
 # ==============================================================================
 import numpy as np
-from tensorflow_model_optimization.python.core.quantization.keras.quantizers import Quantizer
-from typing import List
+
+from typing import List, Callable
 
 from model_compression_toolkit.core.common import BaseNode, Logger
 from model_compression_toolkit.core.common.constants import THRESHOLD, SIGNED, RANGE_MIN, RANGE_MAX
 from model_compression_toolkit.core.common.quantization.quantizers.quantizers_helpers import calculate_delta
 from model_compression_toolkit.core.common.target_platform import QuantizationMethod
-from model_compression_toolkit.exporter.fully_quantized.keras.quantizers.fq_quantizer import FakeQuantQuantizer
-from model_compression_toolkit.exporter.fully_quantized.keras.quantizers.weights_uniform_quantizer import \
-    WeightsUniformQuantizer
 
 # Supporting other quantizer types in the future
+from model_compression_toolkit.exporter.model_wrapper.pytorch.quantizers.fq_quantizer import FakeQuantQuantizer
+from model_compression_toolkit.exporter.model_wrapper.pytorch.quantizers.uniform_weights_quantizer import \
+    UniformWeightsQuantizer
+import torch
+
 SUPPORTED_WEIGHT_QUANTIZER_TYPES = [QuantizationMethod.POWER_OF_TWO,
                                     QuantizationMethod.SYMMETRIC,
                                     QuantizationMethod.UNIFORM]
@@ -34,13 +36,12 @@ SUPPORTED_ACTIVATION_QUANTIZER_TYPES = [QuantizationMethod.POWER_OF_TWO,
                                         QuantizationMethod.UNIFORM]
 
 
-def get_weights_quantizer_for_node(node: BaseNode, weights_attr: List[str]) -> Quantizer:
+def get_weights_quantizer_for_node(node: BaseNode) -> List[Callable]:
     """
     Get weights quantizer for a node.
 
     Args:
         node: Node to create a weight quantizer for.
-        weights_attr: Attributes of the layer to quantize its weights.
 
     Returns:
         Quantizer for the node's weights.
@@ -55,10 +56,10 @@ def get_weights_quantizer_for_node(node: BaseNode, weights_attr: List[str]) -> Q
     if weights_quantization_method not in SUPPORTED_WEIGHT_QUANTIZER_TYPES:
         Logger.error(f'Fully quantized models are now supported for {SUPPORTED_WEIGHT_QUANTIZER_TYPES} quantization methods, but node has {weights_quantization_method} quantization method')
 
-    weight_thresholds = node_w_qc.weights_quantization_params.get(THRESHOLD)
-
     # Compute quantizer params based on node's quantization params
     if weights_quantization_method in [QuantizationMethod.POWER_OF_TWO, QuantizationMethod.SYMMETRIC]:
+        weight_thresholds = node_w_qc.weights_quantization_params.get(THRESHOLD)
+        assert weight_thresholds is not None
         if weights_quantization_method == QuantizationMethod.POWER_OF_TWO:
             is_threshold_pot = np.all([int(np.log2(x)) == np.log2(x) for x in weight_thresholds.flatten()])
             if not is_threshold_pot:
@@ -69,24 +70,19 @@ def get_weights_quantizer_for_node(node: BaseNode, weights_attr: List[str]) -> Q
                                                         n_bits=node_w_qc.weights_n_bits,
                                                         signed=True)
 
-    elif weights_quantization_method in [QuantizationMethod.UNIFORM]:
-        min_range = node_w_qc.weights_quantization_params.get(RANGE_MIN)
-        max_range = node_w_qc.weights_quantization_params.get(RANGE_MAX)
-
     else:
         Logger.error(f'For now fully quantized models support only {SUPPORTED_WEIGHT_QUANTIZER_TYPES} for weights quantization, but found {weights_quantization_method}')
 
-    if len(weights_attr) > 1:
-        Logger.error(f'Currently, we support only one quantized weight per layer, but received {len(weights_attr)} attributes to quantize')
+    return [UniformWeightsQuantizer(num_bits=node_w_qc.weights_n_bits,
+                                    max_range=max_range,
+                                    min_range=min_range,
+                                    quantization_method=node_w_qc.weights_quantization_method,
+                                    per_channel=node_w_qc.weights_per_channel_threshold,
+                                    output_channels_axis=node_w_qc.weights_channels_axis
+                                    )]
 
-    return WeightsUniformQuantizer(nbits=node_w_qc.weights_n_bits,
-                                   min_range=min_range,
-                                   max_range=max_range,
-                                   weight=node.get_weights_by_keys(weights_attr[0]),
-                                   quantization_method=weights_quantization_method)
 
-
-def get_activations_quantizer_for_node(node: BaseNode) -> Quantizer:
+def get_activations_quantizer_for_node(node: BaseNode) -> List[Callable]:
     """
     Get activation quantizer for a node.
 
@@ -125,10 +121,15 @@ def get_activations_quantizer_for_node(node: BaseNode) -> Quantizer:
             activation_thresholds,
             n_bits=node_act_qc.activation_n_bits,
             signed=node_act_qc.activation_quantization_params.get(SIGNED))
+
+    elif activation_quantization_method in [QuantizationMethod.UNIFORM]:
+        min_range = node_act_qc.activation_quantization_params.get(RANGE_MIN)
+        max_range = node_act_qc.activation_quantization_params.get(RANGE_MAX)
+
     else:
         Logger.error(f'For now fully quantized models support only {SUPPORTED_ACTIVATION_QUANTIZER_TYPES} for activation quantization, but found {activation_quantization_method}')
 
-    return FakeQuantQuantizer(nbits=node_act_qc.activation_n_bits,
+    return [FakeQuantQuantizer(nbits=node_act_qc.activation_n_bits,
                               min_range=min_range,
                               max_range=max_range,
-                              quantization_method=activation_quantization_method)
+                              quantization_method=activation_quantization_method)]
