+++ model_compression_toolkit/exporter/model_wrapper/pytorch/quantizers/uniform_weights_quantizer.py
