--- model_compression_toolkit/exporter/target_platform_export/keras/exporters/base_keras_exporter.py
@@ -1,37 +1,0 @@
-# Copyright 2022 Sony Semiconductor Israel, Inc. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-from model_compression_toolkit.exporter.target_platform_export.fw_agonstic.exporter import Exporter
-import keras
-
-class BaseKerasExporter(Exporter):
-    """
-    Base Keras exporter class.
-    """
-
-    def __init__(self, model: keras.models.Model):
-        """
-
-        Args:
-            model: Keras model to export.
-        """
-        super().__init__(model)
-
-    @staticmethod
-    def get_custom_objects(self):
-        return {}
-
-
-
-
