+++ model_compression_toolkit/gptq/pytorch/quantization_facade.py
@@ -38,7 +38,7 @@ if FOUND_TORCH:
     from model_compression_toolkit.core.pytorch.pytorch_implementation import PytorchImplementation
     from model_compression_toolkit.core.pytorch.constants import DEFAULT_TP_MODEL
     from model_compression_toolkit.gptq.pytorch.gptq_loss import multiple_tensors_mse_loss
-    from model_compression_toolkit.exporter.fully_quantized.pytorch.builder.fully_quantized_model_builder import get_fully_quantized_pytorch_model
+    from model_compression_toolkit.exporter.model_wrapper.pytorch.builder.fully_quantized_model_builder import get_fully_quantized_pytorch_model
     import torch
     from torch.nn import Module
     from torch.optim import Adam, Optimizer
